mesa

+++ simple_app.py
@@ -11,9 +11,9 @@ def show_driver_profile(post_id):
     try:
         new_driver = Drivers()
         resp = str(new_driver.show_drivers(post_id))
-        if resp == '[]':
+        if resp == '[]' :
             return Response('Объект не найден в базе', status=404)
-        return resp
+        return str(resp)
     except Exception:
         return Response('Неправильный запрос', status=400)
 
@@ -46,8 +46,14 @@ def driver():
 
 @app.route('/clients/<int:client_id>')
 def show_client_profile(client_id):
-    new_client = Clients()
-    return str(new_client.show_clients(client_id))
+    try:
+        new_client = Clients()
+        resp = str(new_client.show_clients(client_id))
+        if resp == '[]':
+            return Response('Объект не найден в базе', status=404)
+        return resp
+    except Exception:
+        return Response('Неправильный запрос', status=400)
 
 
 @app.route('/clients', methods=['POST', 'DELETE'])
@@ -63,12 +69,17 @@ def client():
             new_client.insert_clients(json_from_request['name'], json_from_request['is_vip'])
             return Response('Created', status=201)
         except Exception:
-            return Response('Произошла ошибка', status=400)
+            return Response('Неправильный запрос', status=400)
     elif request.method == 'DELETE':
-        print(json_from_request)
-        print((json_from_request['id']))
-        new_client.delete_clients(json_from_request['id'])
-        return Response('Deleted', status=201)
+        try:
+            print(json_from_request)
+            print((json_from_request['id']))
+            if str(new_client.show_clients(json_from_request['id'])) == '[]':
+                return Response('Объект не найден в базе', status=404)
+            new_client.delete_clients(json_from_request['id'])
+            return Response('Deleted', status=201)
+        except Exception:
+            return Response('Неправильный запрос', status=400)
 
 
 @app.route('/orders', methods=['GET', 'POST', 'DELETE'])
